Split large class and other improvements
