+++ ts/model/Attribute.ts
@@ -1,0 +1,7 @@
+export class Attribute {
+    public id: string;
+    public dataSourceId: number;
+    public name: string;
+    public value: string;
+    public attributeType: string;
+}
