+++ ts/model/Check.ts
@@ -1,0 +1,5 @@
+export class Check {
+    public id: string;
+    public tenantId: string;
+    public name: string;
+}
