+++ ts/model/IHash.ts
@@ -1,0 +1,3 @@
+export interface IHash<V> {
+    [key: string]: V;
+}
