+++ ts/model/Relation.ts
@@ -1,0 +1,5 @@
+export class Relation {
+    public id: string;
+    public dataSourceId: number;
+    public fqn: string;
+}
