+++ ts/model/Sample.ts
@@ -1,0 +1,10 @@
+export class Sample {
+    public timestamp: Date;
+    public val: number;
+    public min: number;
+    public max: number;
+    public avg: number;
+    public sum: number;
+    public cnt: number;
+    public actual: number;
+}
