+++ ts/model/Tag.ts
@@ -1,0 +1,6 @@
+export class Tag {
+    public id: string;
+    public dataSourceId: number;
+    public name: string;
+    public value: string;
+}
